% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_for_pesgm.R
\name{calc.begin.datetime}
\alias{calc.begin.datetime}
\title{Given DOT begin datetime, calculate the overall begin datetime.}
\usage{
calc.begin.datetime(df, cur.datetime, c = 48, D = 30)
}
\arguments{
\item{df}{the input data frame.}

\item{cur.datetime}{DOT begin datetime.}

\item{c}{number of time intervals per day.}

\item{D}{number of days to look back.}
}
\description{
Given DOT begin datetime, calculate the overall begin datetime.
}
