% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_for_pesgm.R
\name{calc.detect.end}
\alias{calc.detect.end}
\title{Given an anomaly occurrence time, calculate the detection period end time.}
\usage{
calc.detect.end(df, oc.datetime, timeout = 48)
}
\arguments{
\item{df}{the input data frame.}

\item{oc.datetime}{anomaly occurrence time.}

\item{timeout}{maximum acceptable period (number of intervals) for a useful detection.}
}
\description{
Given an anomaly occurrence time, calculate the detection period end time.
}
