% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_for_pesgm.R
\name{calc.DOT}
\alias{calc.DOT}
\title{Compute the D-day same-time avearge for the T-hour sd for each point in range
and do single-point anomaly detection.}
\usage{
calc.DOT(df, range, alpha = 0.9, w = c(0.6, 0.3, 0.1))
}
\arguments{
\item{df}{the input data frame.}

\item{range}{the datetime range for calculation.}

\item{alpha}{confidence level in the Grubb's test.}

\item{w}{weight vector for the three features.}
}
\description{
Compute the D-day same-time avearge for the T-hour sd for each point in range
and do single-point anomaly detection.
}
