% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_for_pesgm.R
\name{calc.T2.datetime}
\alias{calc.T2.datetime}
\title{Given DOT begin datetime, calculate the T2 begin datetime.}
\usage{
calc.T2.datetime(df, cur.datetime, c = 48, D = 30)
}
\arguments{
\item{df}{the input data frame.}

\item{cur.datetime}{the reference datetime.}

\item{c}{number of time intervals per day.}

\item{D}{number of days to look back.}
}
\description{
Given DOT begin datetime, calculate the T2 begin datetime.
}
