% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_for_pesgm.R
\name{calc.T2}
\alias{calc.T2}
\title{Compute the past 24-hour and 6-hour mean and sd for each point in range.}
\usage{
calc.T2(df, range)
}
\arguments{
\item{df}{the input data frame.}

\item{range}{the datetime range for calculation.}
}
\description{
Compute the past 24-hour and 6-hour mean and sd for each point in range.
}
