% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_for_pesgm.R
\name{count_anomaly}
\alias{count_anomaly}
\title{Count number of anomalies within a time range.}
\usage{
count_anomaly(df, range, K = 6)
}
\arguments{
\item{df}{the input data frame.}

\item{range}{the datetime range, a vector of two elements.}

\item{K}{the minimum number of contiguous single-point anomalies needed to conclude an anomalous sequence.}
}
\description{
Count number of anomalies within a time range.
}
