% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_for_pesgm.R
\name{detect.simu}
\alias{detect.simu}
\title{Run an experiment.}
\usage{
detect.simu(df, range, N)
}
\arguments{
\item{df}{the input data frame.}

\item{range}{time range in which simulated anomalies can occur.}

\item{N}{number of detection experiments to run.}
}
\description{
Generate N random cases in range and report detection results.
}
