% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_for_pesgm.R
\name{find_anomaly}
\alias{find_anomaly}
\title{Find the first anomaly after a given time point
and return the number of intervals past until detection.}
\usage{
find_anomaly(df, start.time, K = 6, timeout = 48)
}
\arguments{
\item{df}{the input data frame.}

\item{start.time}{the starting datetime.}

\item{K}{the minimum number of contiguous single-point anomalies needed to conclude an anomalous sequence.}

\item{timeout}{maximum acceptable period (number of intervals) for a useful detection.}
}
\description{
Find the first anomaly after a given time point
and return the number of intervals past until detection.
}
