% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_for_pesgm.R
\name{makeAbCase}
\alias{makeAbCase}
\title{Make an anomaly data case given occurrence time.}
\usage{
makeAbCase(df.in, occurrence.time)
}
\arguments{
\item{df.in}{the input data frame.}

\item{occurrence.time}{time at which the anomaly is to occur.}
}
\description{
Make an anomaly data case given occurrence time.
}
