% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_for_pesgm.R
\name{sample.occurrence}
\alias{sample.occurrence}
\title{Sample N occurrence time within range.}
\usage{
sample.occurrence(df, range, N = 10, seed = 2017)
}
\arguments{
\item{df}{the input data frame.}

\item{range}{the datetime range within which samples are drawn. It is a
vector of two elements, which mark the start and end datetime of the range.}

\item{N}{number of samples to draw.}

\item{seed}{random seed, an integer.}
}
\description{
Sample N occurrence time within range.
}
